+++ Apromore-Custom-Plugins/CSVImporter-Logic/src/main/java/org/apromore/service/csvimporter/services/CSVToParqeutExporter.java
@@ -27,31 +27,26 @@ import com.opencsv.CSVReader;
 import org.apache.commons.io.input.ReaderInputStream;
 import org.apache.hadoop.fs.Path;
 import org.apache.parquet.schema.MessageType;
-import org.apromore.service.csvimporter.dateparser.Parse;
 import org.apromore.service.csvimporter.io.CSVFileReader;
 import org.apromore.service.csvimporter.io.ParquetFileWriter;
 import org.apromore.service.csvimporter.model.*;
 
 import java.io.*;
 import java.nio.charset.Charset;
-import java.sql.Timestamp;
 import java.util.ArrayList;
-import java.util.HashMap;
+import java.util.Arrays;
 import java.util.List;
-import java.util.Map;
 
 import static org.apromore.service.csvimporter.utilities.CSVUtilities.getMaxOccurringChar;
 import static org.apromore.service.csvimporter.utilities.ParquetUtilities.createParquetSchema;
 
 class CSVToParqeutExporter implements ParquetExporter {
-    private final Parse parse = new Parse();
-    boolean preferMonthFirstChanged;
+
     private List<LogErrorReport> logErrorReport;
-    private boolean validRow;
+    private LogProcessor logProcessor;
 
     @Override
     public LogModel generateParqeuetFile(InputStream in, LogSample sample, String charset, File outputParquet, boolean skipInvalidRow) throws Exception {
-
 
         sample.validateSample();
         //If file exist, delete it
@@ -85,29 +80,17 @@ class CSVToParqeutExporter implements ParquetExporter {
             }
         }
 
+        logProcessor = new LogProcessorImpl();
         logErrorReport = new ArrayList<>();
         int lineIndex = 1; // set to 1 since first line is the header
         int numOfValidEvents = 0;
-        boolean preferMonthFirst = preferMonthFirstChanged = parse.getPreferMonthFirst();
-
         String[] line;
-
-        String caseId;
-        String activity;
-        Timestamp endTimestamp;
-        Timestamp startTimestamp;
-        String resource;
-        HashMap<String, String> caseAttributes;
-        HashMap<String, String> eventAttributes;
-        HashMap<String, Timestamp> otherTimestamps;
-
-        String errorMessage = "Field is empty or has a null value!";
         boolean rowLimitExceeded = false;
+        LogEventModelExt logEventModelExt;
 
         while ((line = reader.readNext()) != null && isValidLineCount(lineIndex - 1)) {
 
             // new row, new event.
-            validRow = true;
             lineIndex++;
 
             //empty row
@@ -121,65 +104,10 @@ class CSVToParqeutExporter implements ParquetExporter {
             }
 
             //Construct an event
-            startTimestamp = null;
-            resource = null;
-            caseAttributes = new HashMap<>();
-            eventAttributes = new HashMap<>();
-            otherTimestamps = new HashMap<>();
-
-
-            // Case id:
-            caseId = line[sample.getCaseIdPos()];
-            if (caseId == null || caseId.isEmpty()) {
-                invalidRow(new LogErrorReportImpl(lineIndex, sample.getCaseIdPos(), header[sample.getCaseIdPos()], errorMessage));
-            }
-
-            // Activity
-            activity = line[sample.getActivityPos()];
-            if (activity == null || activity.isEmpty()) {
-                invalidRow(new LogErrorReportImpl(lineIndex, sample.getActivityPos(), header[sample.getActivityPos()], errorMessage));
-            }
-
-            // End Timestamp
-            endTimestamp = parseTimestampValue(line[sample.getEndTimestampPos()], sample.getEndTimestampFormat());
-            if (endTimestamp == null) {
-                invalidRow(new LogErrorReportImpl(lineIndex, sample.getEndTimestampPos(), header[sample.getEndTimestampPos()], parse.getParseFailMess()));
-            }
-            // Start Timestamp
-            if (sample.getStartTimestampPos() != -1) {
-                startTimestamp = parseTimestampValue(line[sample.getStartTimestampPos()], sample.getStartTimestampFormat());
-                if (startTimestamp == null) {
-                    invalidRow(new LogErrorReportImpl(lineIndex, sample.getStartTimestampPos(), header[sample.getStartTimestampPos()], parse.getParseFailMess()));
-                }
-            }
-
-            // Other timestamps
-            if (!sample.getOtherTimestamps().isEmpty()) {
-                for (Map.Entry<Integer, String> otherTimestamp : sample.getOtherTimestamps().entrySet()) {
-                    Timestamp tempTimestamp = parseTimestampValue(line[otherTimestamp.getKey()], otherTimestamp.getValue());
-                    if (tempTimestamp != null) {
-                        otherTimestamps.put(header[otherTimestamp.getKey()], tempTimestamp);
-                    } else {
-                        invalidRow(new LogErrorReportImpl(lineIndex, otherTimestamp.getKey(), header[otherTimestamp.getKey()], parse.getParseFailMess()));
-                    }
-                }
-            }
-
-            // If PreferMonthFirst changed to True, we have to start over.
-            if (!preferMonthFirst && preferMonthFirstChanged)
-                generateParqeuetFile(in, sample, charset, outputParquet, skipInvalidRow);
-
-
-            // Resource
-            if (sample.getResourcePos() != -1) {
-                resource = line[sample.getResourcePos()];
-                if (resource == null || resource.isEmpty()) {
-                    invalidRow(new LogErrorReportImpl(lineIndex, sample.getResourcePos(), header[sample.getResourcePos()], errorMessage));
-                }
-            }
+            logEventModelExt = logProcessor.processLog(Arrays.asList(line), Arrays.asList(header), sample, lineIndex, logErrorReport);
 
             // If row is invalid, continue to next row.
-            if (!validRow) {
+            if (!logEventModelExt.isValid()) {
                 if (skipInvalidRow) {
                     continue;
                 } else {
@@ -187,21 +115,7 @@ class CSVToParqeutExporter implements ParquetExporter {
                 }
             }
 
-            // Case Attributes
-            if (sample.getCaseAttributesPos() != null && !sample.getCaseAttributesPos().isEmpty()) {
-                for (int columnPos : sample.getCaseAttributesPos()) {
-                    caseAttributes.put(header[columnPos], line[columnPos]);
-                }
-            }
-
-            // Event Attributes
-            if (sample.getEventAttributesPos() != null && !sample.getEventAttributesPos().isEmpty()) {
-                for (int columnPos : sample.getEventAttributesPos()) {
-                    eventAttributes.put(header[columnPos], line[columnPos]);
-                }
-            }
-
-            writer.write(new LogEventModel(caseId, activity, endTimestamp, startTimestamp, otherTimestamps, resource, eventAttributes, caseAttributes));
+            writer.write(logEventModelExt);
             numOfValidEvents++;
         }
         writer.close();
@@ -214,29 +128,5 @@ class CSVToParqeutExporter implements ParquetExporter {
 
     public boolean isValidLineCount(int lineCount) {
         return true;
-    }
-
-    private Timestamp parseTimestampValue(String theValue, String format) {
-        Timestamp stamp;
-        if (format != null && !format.isEmpty()) {
-            stamp = parse.tryParsingWithFormat(theValue, format);
-            if (stamp == null) {
-                stamp = parse.tryParsing(theValue);
-                if (!preferMonthFirstChanged) {
-                    preferMonthFirstChanged = parse.getPreferMonthFirst();
-                }
-            }
-        } else {
-            stamp = parse.tryParsing(theValue);
-            if (!preferMonthFirstChanged) {
-                preferMonthFirstChanged = parse.getPreferMonthFirst();
-            }
-        }
-        return stamp;
-    }
-
-    private void invalidRow(LogErrorReportImpl error) {
-        logErrorReport.add(error);
-        validRow = false;
     }
 }
