+++ Apromore-Custom-Plugins/CSVImporter-Logic/src/main/java/org/apromore/service/csvimporter/services/LogProcessorImpl.java
@@ -96,7 +96,6 @@ public class LogProcessorImpl implements LogProcessor {
             }
         }
 
-
         // Resource
         if (sample.getResourcePos() != -1) {
             resource = line.get(sample.getResourcePos());
