+++ Apromore-Custom-Plugins/CSVImporter-Logic/src/main/java/org/apromore/service/csvimporter/services/ParquetToParquetExporter.java
@@ -25,7 +25,6 @@ import org.apache.hadoop.conf.Configuration;
 import org.apache.parquet.example.data.Group;
 import org.apache.parquet.hadoop.ParquetReader;
 import org.apache.parquet.schema.MessageType;
-import org.apromore.service.csvimporter.dateparser.Parse;
 import org.apromore.service.csvimporter.io.ParquetFileWriter;
 import org.apromore.service.csvimporter.io.ParquetLocalFileReader;
 import org.apromore.service.csvimporter.io.ParquetLocalFileWriter;
@@ -33,21 +32,17 @@ import org.apromore.service.csvimporter.model.*;
 
 import java.io.File;
 import java.io.InputStream;
-import java.io.UnsupportedEncodingException;
-import java.sql.Timestamp;
 import java.util.ArrayList;
-import java.util.HashMap;
+import java.util.Arrays;
 import java.util.List;
-import java.util.Map;
 
 import static org.apromore.service.csvimporter.utilities.ParquetUtilities.createParquetSchema;
 import static org.apromore.service.csvimporter.utilities.ParquetUtilities.getHeaderFromParquet;
 
 class ParquetToParquetExporter implements ParquetExporter {
-    private final Parse parse = new Parse();
-    boolean preferMonthFirstChanged;
+
     private List<LogErrorReport> logErrorReport;
-    private boolean validRow;
+    private LogProcessor logProcessor;
 
     @Override
     public LogModel generateParqeuetFile(InputStream in, LogSample sample, String charset, File outputParquet, boolean skipInvalidRow) throws Exception {
@@ -75,24 +70,13 @@ class ParquetToParquetExporter implements ParquetExporter {
         MessageType sampleSchema = createParquetSchema(header, sample);
         ParquetFileWriter writer = new ParquetLocalFileWriter().getParquetWriter(outputParquet, sampleSchema);
 
+        logProcessor = new LogProcessorImpl();
         logErrorReport = new ArrayList<>();
         int lineIndex = 0;
         int numOfValidEvents = 0;
-        boolean preferMonthFirst = preferMonthFirstChanged = parse.getPreferMonthFirst();
-
         String[] line;
-
-        String caseId;
-        String activity;
-        Timestamp endTimestamp;
-        Timestamp startTimestamp;
-        String resource;
-        HashMap<String, String> caseAttributes;
-        HashMap<String, String> eventAttributes;
-        HashMap<String, Timestamp> otherTimestamps;
-
-        String errorMessage = "Field is empty or has a null value!";
         boolean rowLimitExceeded = false;
+        LogEventModelExt logEventModelExt;
 
         Group g;
         while ((g = reader.read()) != null && isValidLineCount(lineIndex)) {
@@ -105,7 +89,6 @@ class ParquetToParquetExporter implements ParquetExporter {
             }
 
             // new row, new event.
-            validRow = true;
             lineIndex++;
 
             //empty row
@@ -119,65 +102,10 @@ class ParquetToParquetExporter implements ParquetExporter {
             }
 
             //Construct an event
-            startTimestamp = null;
-            resource = null;
-            caseAttributes = new HashMap<>();
-            eventAttributes = new HashMap<>();
-            otherTimestamps = new HashMap<>();
-
-
-            // Case id:
-            caseId = line[sample.getCaseIdPos()];
-            if (caseId == null || caseId.isEmpty()) {
-                invalidRow(new LogErrorReportImpl(lineIndex, sample.getCaseIdPos(), header[sample.getCaseIdPos()], errorMessage));
-            }
-
-            // Activity
-            activity = line[sample.getActivityPos()];
-            if (activity == null || activity.isEmpty()) {
-                invalidRow(new LogErrorReportImpl(lineIndex, sample.getActivityPos(), header[sample.getActivityPos()], errorMessage));
-            }
-
-            // End Timestamp
-            endTimestamp = parseTimestampValue(line[sample.getEndTimestampPos()], sample.getEndTimestampFormat());
-            if (endTimestamp == null) {
-                invalidRow(new LogErrorReportImpl(lineIndex, sample.getEndTimestampPos(), header[sample.getEndTimestampPos()], parse.getParseFailMess()));
-            }
-            // Start Timestamp
-            if (sample.getStartTimestampPos() != -1) {
-                startTimestamp = parseTimestampValue(line[sample.getStartTimestampPos()], sample.getStartTimestampFormat());
-                if (startTimestamp == null) {
-                    invalidRow(new LogErrorReportImpl(lineIndex, sample.getStartTimestampPos(), header[sample.getStartTimestampPos()], parse.getParseFailMess()));
-                }
-            }
-
-            // Other timestamps
-            if (!sample.getOtherTimestamps().isEmpty()) {
-                for (Map.Entry<Integer, String> otherTimestamp : sample.getOtherTimestamps().entrySet()) {
-                    Timestamp tempTimestamp = parseTimestampValue(line[otherTimestamp.getKey()], otherTimestamp.getValue());
-                    if (tempTimestamp != null) {
-                        otherTimestamps.put(header[otherTimestamp.getKey()], tempTimestamp);
-                    } else {
-                        invalidRow(new LogErrorReportImpl(lineIndex, otherTimestamp.getKey(), header[otherTimestamp.getKey()], parse.getParseFailMess()));
-                    }
-                }
-            }
-
-            // If PreferMonthFirst changed to True, we have to start over.
-            if (!preferMonthFirst && preferMonthFirstChanged)
-                generateParqeuetFile(in, sample, charset, outputParquet, skipInvalidRow);
-
-
-            // Resource
-            if (sample.getResourcePos() != -1) {
-                resource = line[sample.getResourcePos()];
-                if (resource == null || resource.isEmpty()) {
-                    invalidRow(new LogErrorReportImpl(lineIndex, sample.getResourcePos(), header[sample.getResourcePos()], errorMessage));
-                }
-            }
+            logEventModelExt = logProcessor.processLog(Arrays.asList(line), Arrays.asList(header), sample, lineIndex, logErrorReport);
 
             // If row is invalid, continue to next row.
-            if (!validRow) {
+            if (!logEventModelExt.isValid()) {
                 if (skipInvalidRow) {
                     continue;
                 } else {
@@ -186,27 +114,14 @@ class ParquetToParquetExporter implements ParquetExporter {
                 }
             }
 
-            // Case Attributes
-            if (sample.getCaseAttributesPos() != null && !sample.getCaseAttributesPos().isEmpty()) {
-                for (int columnPos : sample.getCaseAttributesPos()) {
-                    caseAttributes.put(header[columnPos], line[columnPos]);
-                }
-            }
-
-            // Event Attributes
-            if (sample.getEventAttributesPos() != null && !sample.getEventAttributesPos().isEmpty()) {
-                for (int columnPos : sample.getEventAttributesPos()) {
-                    eventAttributes.put(header[columnPos], line[columnPos]);
-                }
-            }
-
-            writer.write(new LogEventModel(caseId, activity, endTimestamp, startTimestamp, otherTimestamps, resource, eventAttributes, caseAttributes));
+            writer.write(logEventModelExt);
             numOfValidEvents++;
         }
         writer.close();
 
         if (!isValidLineCount(lineIndex))
             rowLimitExceeded = true;
+
         //Upon migrating to parquet, xlog need to be removed and LogModelXLogImpl need to be renamed
         return new LogModelXLogImpl(null, logErrorReport, rowLimitExceeded, numOfValidEvents);
     }
@@ -215,31 +130,7 @@ class ParquetToParquetExporter implements ParquetExporter {
         return true;
     }
 
-    private Timestamp parseTimestampValue(String theValue, String format) {
-        Timestamp stamp;
-        if (format != null && !format.isEmpty()) {
-            stamp = parse.tryParsingWithFormat(theValue, format);
-            if (stamp == null) {
-                stamp = parse.tryParsing(theValue);
-                if (!preferMonthFirstChanged) {
-                    preferMonthFirstChanged = parse.getPreferMonthFirst();
-                }
-            }
-        } else {
-            stamp = parse.tryParsing(theValue);
-            if (!preferMonthFirstChanged) {
-                preferMonthFirstChanged = parse.getPreferMonthFirst();
-            }
-        }
-        return stamp;
-    }
-
-    private void invalidRow(LogErrorReportImpl error) {
-        logErrorReport.add(error);
-        validRow = false;
-    }
-
-    private String[] readGroup(Group g, MessageType schema) throws UnsupportedEncodingException {
+    private String[] readGroup(Group g, MessageType schema) {
 
         String[] line = new String[schema.getColumns().size()];
         for (int j = 0; j < schema.getFieldCount(); j++) {
