+++ Apromore-Custom-Plugins/CSVImporter-Logic/src/main/java/org/apromore/service/csvimporter/services/XLSToParquetExporterLegecy.java
@@ -1,0 +1,259 @@
+/*-
+ * #%L
+ * This file is part of "Apromore Core".
+ * %%
+ * Copyright (C) 2018 - 2020 Apromore Pty Ltd.
+ * %%
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Lesser Public License for more details.
+ *
+ * You should have received a copy of the GNU General Lesser Public
+ * License along with this program.  If not, see
+ * <http://www.gnu.org/licenses/lgpl-3.0.html>.
+ * #L%
+ */
+package org.apromore.service.csvimporter.services;
+
+import org.apache.hadoop.fs.Path;
+import org.apache.parquet.schema.MessageType;
+import org.apache.poi.ss.usermodel.Cell;
+import org.apache.poi.ss.usermodel.Row;
+import org.apache.poi.ss.usermodel.Sheet;
+import org.apache.poi.ss.usermodel.Workbook;
+import org.apromore.service.csvimporter.dateparser.Parse;
+import org.apromore.service.csvimporter.io.ParquetFileWriter;
+import org.apromore.service.csvimporter.io.XLSReader;
+import org.apromore.service.csvimporter.model.*;
+
+import java.io.File;
+import java.io.InputStream;
+import java.sql.Timestamp;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+import static org.apromore.service.csvimporter.utilities.ParquetUtilities.createParquetSchema;
+
+public class XLSToParquetExporterLegecy implements ParquetExporter {
+
+    private final Parse parse = new Parse();
+    boolean preferMonthFirstChanged;
+    private List<LogErrorReport> logErrorReport;
+    private boolean validRow;
+    private final int BUFFER_SIZE = 2048;
+    private final int DEFAULT_NUMBER_OF_ROWS = 100;
+
+    @Override
+    public LogModel generateParqeuetFile(InputStream in, LogSample sample, String charset, File outputParquet, boolean skipInvalidRow) throws Exception {
+
+
+        sample.validateSample();
+        //If file exist, delete it
+        if (outputParquet.exists())
+            outputParquet.delete();
+
+        Workbook workbook = new XLSReader().readXLS(in, DEFAULT_NUMBER_OF_ROWS, BUFFER_SIZE);
+        if (workbook == null)
+            throw new Exception("Unable to import file");
+
+        List<String> headerList = new ArrayList<>();
+        Sheet sheet = workbook.getSheetAt(0);
+
+        //Get the header
+        if (sheet == null)
+            throw new Exception("Unable to import file");
+
+        for (Row r : sheet) {
+            for (Cell c : r) {
+                headerList.add(c.getStringCellValue());
+            }
+            break;
+        }
+        String[] header = headerList.toArray(new String[0]);
+
+        MessageType parquetSchema = createParquetSchema(header, sample);
+        ParquetFileWriter writer;
+
+        // Classpath manipulation so that ServiceLoader in parquet-osgi reads its own META-INF/services rather than the servlet context bundle's (i.e. the portal)
+        Thread thread = Thread.currentThread();
+        synchronized (thread) {
+            ClassLoader originalContextClassLoader = thread.getContextClassLoader();
+            try {
+                thread.setContextClassLoader(Path.class.getClassLoader());
+                writer = new ParquetFileWriter(new Path(outputParquet.toURI()), parquetSchema, true);
+            } finally {
+                thread.setContextClassLoader(originalContextClassLoader);
+            }
+        }
+
+        logErrorReport = new ArrayList<>();
+        int lineIndex = 1; // set to 1 since first line is the header
+        int numOfValidEvents = 0;
+        boolean preferMonthFirst = preferMonthFirstChanged = parse.getPreferMonthFirst();
+
+        ArrayList<String> line;
+
+        String caseId;
+        String activity;
+        Timestamp endTimestamp;
+        Timestamp startTimestamp;
+        String resource;
+        HashMap<String, String> caseAttributes;
+        HashMap<String, String> eventAttributes;
+        HashMap<String, Timestamp> otherTimestamps;
+
+        String errorMessage = "Field is empty or has a null value!";
+        boolean rowLimitExceeded = false;
+        for (Row r : sheet) {
+
+            if (!isValidLineCount(lineIndex - 1))
+                break;
+
+            // new row, new event.
+            validRow = true;
+            lineIndex++;
+            line = new ArrayList<>();
+
+            //Get the rows
+            for (Cell c : r) {
+                line.add(c.getStringCellValue());
+            }
+
+            //empty row
+            if (line.size() == 0 || (line.size() == 1 && (line.get(0).trim().equals("") || line.get(0).trim().equals("\n"))))
+                continue;
+
+            //Validate num of column
+            if (header.length != line.size()) {
+                logErrorReport.add(new LogErrorReportImpl(lineIndex, 0, null, "Number of columns does not match the number of headers. Number of headers: (" + header.length + "). Number of columns: (" + line.size() + ")"));
+                continue;
+            }
+
+            //Construct an event
+            startTimestamp = null;
+            resource = null;
+            caseAttributes = new HashMap<>();
+            eventAttributes = new HashMap<>();
+            otherTimestamps = new HashMap<>();
+
+
+            // Case id:
+            caseId = line.get(sample.getCaseIdPos());
+            if (caseId == null || caseId.isEmpty()) {
+                invalidRow(new LogErrorReportImpl(lineIndex, sample.getCaseIdPos(), header[sample.getCaseIdPos()], errorMessage));
+            }
+
+            // Activity
+            activity = line.get(sample.getActivityPos());
+            if (activity == null || activity.isEmpty()) {
+                invalidRow(new LogErrorReportImpl(lineIndex, sample.getActivityPos(), header[sample.getActivityPos()], errorMessage));
+            }
+
+            // End Timestamp
+            endTimestamp = parseTimestampValue(line.get(sample.getEndTimestampPos()), sample.getEndTimestampFormat());
+            if (endTimestamp == null) {
+                invalidRow(new LogErrorReportImpl(lineIndex, sample.getEndTimestampPos(), header[sample.getEndTimestampPos()], parse.getParseFailMess()));
+            }
+            // Start Timestamp
+            if (sample.getStartTimestampPos() != -1) {
+                startTimestamp = parseTimestampValue(line.get(sample.getStartTimestampPos()), sample.getStartTimestampFormat());
+                if (startTimestamp == null) {
+                    invalidRow(new LogErrorReportImpl(lineIndex, sample.getStartTimestampPos(), header[sample.getStartTimestampPos()], parse.getParseFailMess()));
+                }
+            }
+
+            // Other timestamps
+            if (!sample.getOtherTimestamps().isEmpty()) {
+                for (Map.Entry<Integer, String> otherTimestamp : sample.getOtherTimestamps().entrySet()) {
+                    Timestamp tempTimestamp = parseTimestampValue(line.get(otherTimestamp.getKey()), otherTimestamp.getValue());
+                    if (tempTimestamp != null) {
+                        otherTimestamps.put(header[otherTimestamp.getKey()], tempTimestamp);
+                    } else {
+                        invalidRow(new LogErrorReportImpl(lineIndex, otherTimestamp.getKey(), header[otherTimestamp.getKey()], parse.getParseFailMess()));
+                    }
+                }
+            }
+
+            // If PreferMonthFirst changed to True, we have to start over.
+            if (!preferMonthFirst && preferMonthFirstChanged)
+                generateParqeuetFile(in, sample, charset, outputParquet, skipInvalidRow);
+
+
+            // Resource
+            if (sample.getResourcePos() != -1) {
+                resource = line.get(sample.getResourcePos());
+                if (resource == null || resource.isEmpty()) {
+                    invalidRow(new LogErrorReportImpl(lineIndex, sample.getResourcePos(), header[sample.getResourcePos()], errorMessage));
+                }
+            }
+
+            // If row is invalid, continue to next row.
+            if (!validRow) {
+                if (skipInvalidRow) {
+                    continue;
+                } else {
+                    return new LogModelXLogImpl(null, logErrorReport, rowLimitExceeded, numOfValidEvents);
+                }
+            }
+
+            // Case Attributes
+            if (sample.getCaseAttributesPos() != null && !sample.getCaseAttributesPos().isEmpty()) {
+                for (int columnPos : sample.getCaseAttributesPos()) {
+                    caseAttributes.put(header[columnPos], line.get(columnPos));
+                }
+            }
+
+            // Event Attributes
+            if (sample.getEventAttributesPos() != null && !sample.getEventAttributesPos().isEmpty()) {
+                for (int columnPos : sample.getEventAttributesPos()) {
+                    eventAttributes.put(header[columnPos], line.get(columnPos));
+                }
+            }
+
+            writer.write(new LogEventModel(caseId, activity, endTimestamp, startTimestamp, otherTimestamps, resource, eventAttributes, caseAttributes));
+            numOfValidEvents++;
+        }
+        writer.close();
+
+        if (!isValidLineCount(lineIndex - 1))
+            rowLimitExceeded = true;
+
+        return new LogModelXLogImpl(null, logErrorReport, rowLimitExceeded, numOfValidEvents);
+    }
+
+    public boolean isValidLineCount(int lineCount) {
+        return true;
+    }
+
+    private Timestamp parseTimestampValue(String theValue, String format) {
+        Timestamp stamp;
+        if (format != null && !format.isEmpty()) {
+            stamp = parse.tryParsingWithFormat(theValue, format);
+            if (stamp == null) {
+                stamp = parse.tryParsing(theValue);
+                if (!preferMonthFirstChanged) {
+                    preferMonthFirstChanged = parse.getPreferMonthFirst();
+                }
+            }
+        } else {
+            stamp = parse.tryParsing(theValue);
+            if (!preferMonthFirstChanged) {
+                preferMonthFirstChanged = parse.getPreferMonthFirst();
+            }
+        }
+        return stamp;
+    }
+
+    private void invalidRow(LogErrorReportImpl error) {
+        logErrorReport.add(error);
+        validRow = false;
+    }
+}
